Forget Me Not 🌺
A smart, contextualized todo list. Designed to live in your terminal, and to be in your face.

This is just a proof of concept, the only thing it does is printing all the tasks found traversing the revers tree of folders form the directory of execution up to the root (/).